+++ src/models/users.js
@@ -2,34 +2,14 @@ const apiError = require('../errors/apiError');
 
 class User {
 
-    static create(client, data) {
-        
-        return new Promise(function (resolve, reject) {
-            client.query('INSERT INTO users (first_name, last_name, username, email, user_password) VALUES ($1, $2, $3, $4, $5) RETURNING userid', [data.firstName, data.lastName, data.username, data.email, data.password]).then((res) => {
-
-                this.id = res.rows[0].userid;
-                this.firstName = data.firstName;
-                this.lastName = data.lastName;
-                this.username = data.username;
-                this.email = data.email;
-                this.password = data.password;
-
-                resolve(this);
-
-            }).catch((err) => {
-                console.log(err)
-                reject(new apiError(parseInt(err.code), err.detail));
-            })
-        })
+    constructor(userid, firstName, lastName, username, email) {
+        this.id = userid;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.username = username;
+        this.email = email;
+        this.password
     }
-
-    // constructor(userid, firstName, lastName, username, email) {
-    //     this.id = userid;
-    //     this.firstName = firstName;
-    //     this.lastName = lastName;
-    //     this.username = username;
-    //     this.email = email;
-    // }
 
     output() {
         return {
@@ -41,16 +21,6 @@ class User {
                 "email": this.email
             }
         }
-    }
-
-
-
-    retrieve(client) {
-        return new Promise(function (resolve, reject) {
-            client.query('SELECT * FROM users WHERE userid = $1', [this.id], function (err, response) {
-                resolve(response.rows)
-            })
-        })
     }
 
     update(client, data) {
@@ -71,62 +41,35 @@ class User {
 
 }
 
-// // create a user
-// function create(client, data) {
-//     return new Promise(function (resolve, reject) {
-//         client.query('INSERT INTO users (first_name, last_name, username, email, user_password) VALUES ($1, $2, $3, $4, $5) RETURNING userid', [data.firstName, data.lastName, data.username, data.email, data.password]).then((res) => {
+function create(client, data) {
 
-//             var user = new User(res.rows[0].userid, data.firstName, data.lastName, data.username, data.email)
+    return new Promise(function (resolve, reject) {
+        client.query('INSERT INTO users (first_name, last_name, username, email, user_password) VALUES ($1, $2, $3, $4, $5) RETURNING userid', [data.firstName, data.lastName, data.username, data.email, data.password]).then((res) => {
 
-//             resolve(user);
+            var user = new User(res.rows[0].userid, data.firstName, data.lastName, data.username, data.email, data.password)
 
-//         }).catch((err) => {
-//             reject(new apiError(parseInt(err.code), err.detail));
-//         })
-//     })
-// }
+            resolve(user);
 
-// return user with specific id
-// function retrieve(client, id) {
-//     return new Promise(function (resolve, reject) {
-//         client.query('SELECT * FROM users WHERE userid = $1', [id], function (err, response) {
-//             resolve(response.rows)
-//         })
-//     })
-// }
+        }).catch((err) => {
+            console.log(err)
+            reject(new apiError(parseInt(err.code), err.detail));
+        })
+    })
+}
 
-// update user with specific id
-// function update(client, data) {
-//     client.query('UPDATE users SET first_name = $1, last_name = $2, username = $3, email = $4, user_password = $5 WHERE userid = $6', [data.firstName, data.lastName, data.username, data.email, data.password, data.userid])
-// }
+function retrieve(client, id) {
+    return new Promise(function (resolve, reject) {
+        client.query('SELECT * FROM users WHERE userid = $1', [id], function (err, response) {
 
-// return all rows
-// function all(client) {
-//     return new Promise(function (resolve, reject) {
-//         client.query('SELECT * FROM users', function (err, response) {
-//             resolve(response.rows)
-//         })
-//     })
-// }
+            var data = response.rows[0]
+            var user = new User(id, data.first_name, data.last_name, data.username, data.email, data.password)
 
-// function to delete row (cant use delete as function name because it is a reseverd word in JS)
-// function remove(client, id) {
-//     client.query('DELETE FROM users WHERE userid = $1', [id]);
-// }
+            console.log(user);
 
-// forms a data object for the frontend
-// function format(data) {
-//     var frontendData = {};
-
-//     // decode data and form frontend data obj
-//     frontendData.userId = data.userid;
-//     frontendData.firstName = data.first_name;
-//     frontendData.lastName = data.last_name;
-//     frontendData.username = data.username;
-//     frontendData.email = data.email;
-
-//     return frontendData;
-// }
+            resolve(user)
+        })
+    })
+}
 
 function validate(data) {
 
@@ -155,5 +98,10 @@ function validate(data) {
 
 }
 
-module.exports = User;
-module.exports.validate = validate;
+module.exports = {
+    User,
+    "validate": validate,
+    "retrieve": retrieve,
+    "create": create
+}
+